requirements.txt
biopython==1.85 numpy==1.26.4
